quizout-client/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asyncio==3.4.3
attrs==25.3.0
bidict==0.23.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
jsonpickle~=4.1.1
multidict==6.5.0
propcache==0.3.2
PySide6~=6.8.0
PySide6_Addons~=6.8.0
PySide6_Essentials~=6.8.0
python-engineio==4.12.2
python-socketio==5.13.0
shiboken6~=6.8.0
simple-websocket==1.1.0
wsproto==1.2.0
yarl==1.20.1
